.claude/worktrees/nostalgic-curie-0215c6/pyproject.toml
[project]
name = "EventTracker"
version = "0.1.0"
description = "Local-first personal timeline and journal app built with FastAPI and SQLite."
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "beautifulsoup4==4.14.3",
    "fastapi==0.135.3",
    "github-copilot-sdk",
    "httpx==0.28.1",
    "jinja2==3.1.6",
    "markdown>=3.10.2",
    "markitdown[pdf,docx,xlsx,xls,pptx,outlook,audio-transcription]>=0.1.5",
    "youtube-transcript-api>=1.2.4",
    "openai>=2.31.0,<3.0.0",
    "python-dotenv==1.2.2",
    "python-multipart==0.0.26",
    "sqlite-vec==0.1.9",
    "uvicorn[standard]==0.44.0",
]

[dependency-groups]
dev = [
    "playwright>=1.52,<2",
    "pyright>=1.1.404,<2",
    "pytest>=9.0.3,<10",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
include = [
    "app/db.py",
    "app/env.py",
    "app/main.py",
    "app/models.py",
    "app/schemas.py",
    "app/services/entries.py",
    "app/services/ai_generate.py",
    "app/services/embeddings.py",
    "app/services/group_web_search.py",
    "app/services/copilot_runtime.py",
    "app/services/event_chat.py",
    "app/services/search.py",
    "scripts/run_dev.py",
    "scripts/import_entries.py",
    "scripts/init_db.py",
    "tests/test_ai_generate.py",
    "tests/test_db.py",
    "tests/test_group_web_search.py",
    "tests/test_entries.py",
    "tests/test_event_chat.py",
    "tests/test_init_db.py",
    "tests/test_import_entries.py",
    "tests/test_run_dev.py",
    "tests/test_search.py",
    "tests/test_smoke.py",
]
reportMissingTypeStubs = false

